+++ src/main/java/ehb/finalwork/manager/model/Beacon.java
@@ -6,13 +6,13 @@ public class Beacon {
     private String room_id;
     private String name;
     private String description;
-    private Float calibrationFactor;
+    private Double calibrationFactor;
     private Vec3d location;
 
     public Beacon() {
     }
 
-    public Beacon(String room_id, String name, String description, Float calibrationFactor, Vec3d location) {
+    public Beacon(String room_id, String name, String description, Double calibrationFactor, Vec3d location) {
         this.room_id = room_id;
         this.name = name;
         this.description = description;
@@ -44,11 +44,11 @@ public class Beacon {
         this.description = description;
     }
 
-    public Float getCalibrationFactor() {
+    public Double getCalibrationFactor() {
         return calibrationFactor;
     }
 
-    public void setCalibrationFactor(Float calibrationFactor) {
+    public void setCalibrationFactor(Double calibrationFactor) {
         this.calibrationFactor = calibrationFactor;
     }
